Package Refactoring
to keep the same structure as the previous version, "statsPanel" has been moved to the interface layer. We also found a problem with the "testNullability" method of the ExtractMethodProcessor class (com.intellij.refactoring): the class method cannot be found.
Error found in:
- EagerTestStrategy.java
- GeneralFixtureStrategy.java
- LackOfCohesionStrategy.java
