+++ src/main/java/stats/Session.java
@@ -39,9 +39,7 @@ public class Session {
     return kind;
   }
 
-  public void setKind(String kind) {
-    this.kind = kind;
-  }
+  public void setKind(String kind) { this.kind = kind; }
 
   public String getID() {
     return ID;
